Added search class
