+++ src/client/js/page-init.js
@@ -18,10 +18,12 @@
  */
 import {instance as lazyImg} from './lazy-img.js';
 import highlightCategory from './highlight-category.js';
+import {search} from './search-class';
 
 function initializePage() {
   lazyImg.loadImages();
   highlightCategory();
+  search.addHandlers();
   const metaTitleEl = document.querySelector('#meta-title');
   document.title = metaTitleEl ? metaTitleEl.textContent : document.title;
 }
