+++ src/client/js/product.js
@@ -21,7 +21,6 @@ import pageInit from './page-init';
 
 function initializeProductPage() {
   pageInit();
-
   // TODO: Add Code dedicated to Product page only.
 }
