+++ src/data/categories.json
@@ -1,1 +1,1 @@
-["accessories","apparel","bags","drinkware","office","brand"]
+["accessories", "apparel", "bags", "drinkware", "office", "brand"]
