+++ src/server/routes/index.js
@@ -30,7 +30,6 @@ function index(req, res) {
     layout: req.query.fragment ? 'fragment' : 'layout',
     scripts: [
       '/js/home_main.js',
-      '/js/search.js',
     ],
     title: 'PWA Shop',
   });
