+++ src/server/routes/payment.js
@@ -23,7 +23,6 @@ const payment = {
       cart: req.session.cart,
       layout: req.query.fragment ? 'fragment' : 'layout',
       scripts: [
-        '/js/search.js',
       ],
       title: 'PWA Shop: Payment',
     });
