+++ src/templates/views/search.hbs
@@ -1,3 +1,3 @@
-<main id="search">
-  <p>Search!</p>
+<main id="search"  data-category="{{query}}">
+  <h1>Search results for <em>{{query}}<em></h1>
 </main>
